requirements.txt
streamlit
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
pandas
plotly